WIP - changed madpack.py & SQLCommon.m4_in
Added a new m4 macro to SQLCommon.m4_in, so we can use it to find
the absolute path to a sql test setup file, which can be included
by any sql test file.

In madpack.py, we now ignore any files in the test dir with .setup
in the name.