All syntax errors fixed, but design doesn't quite work
Sometimes the INSERT that updates model_output table
creates a duplicate row with the same mst key as exists
on another segment.  I think we didn't worry about this
problem because we assumed that particular mst_keys will
always be written to the same segment.  I think that
assumption is false because they hop from segment to
segment!  I think we can fix this by using DISTRIBUTED BY
mst_key.