Only one external model_output table per host now
Removed seg_id and dist_key from the external model_output
table (aka weights_to_update).  Distribution policy is now
based on mst key instead of dist key.  It has to be mst
key so that we don't end up trying to replace a row on
one host with a row on another host.  Internal tables can
handle an update like this by redistributing, but we
don't have that option so we MUST have it distributed this
way.