Fix AutoML merge issues
automl was referencing internal variables of the FitMultiple class for
the tables names.  But these tables names are chosen by the AutoML
so it should be referencing its own class variables.  The variable
names inside FitMultiple should not be relied upon by anything outside,
otherwise refactors like this end up requiring changes to AutoML.
These changes fix AutoML so that dev-check passes, and remove any
reliance on internal FitMultiple variable names.

There was an additional issue regarding the schema for model_output_info_table.
It is defined both in FitMultiple and in AutoML, but the order of the
columns changed slightly in FitMultiple.  This was unintentional, but
ideally should not have broken AutoML either.  For now, order change
in FitMultiple has been reverted--but in the future we should have
both of them reference a common schema so that changes don't need
to be updated in two different places.