DL: Add tests for model hopper
This is a set of "unit tests" implemented as a dev-check test, since
 they need DB access to work.

 Functions unit tested in deep_learning/tests/madlib_keras_fit_multiple.sql_in:
    FitMultiple.init_schedule()
    FitMultiple.rotate_schedule()
    FitMultiple.run_training()

FitMutiple methods we don't have unit tests for yet:

    populate_object_map()
    init_model_output_tbl()
    init_model_info_tbl()
        load_warm_start_model()
        load_xfer_learning_weights()

Any failures in these should show up pretty quickly in other tests we
run routinely (dev-check, end-to-end performance tests, etc.) so unit
tests are less necessary for those.  But we could choose to add them
at some point.

- Tests schedule creation and rotation
- Tests that models are hopping to right segments in right order

- Tests caching, # msts > segs, = , and # mst < segs
  Uses simulated segments instead of actual segments
  so the same test can be run on any size cluster.

- 6 models on 3 segments
- 3 models on 3 segments
- 3 models on 5 segments